# specs/001-agent-core-spec/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Java Shannon Agent Orchestrator Core API
  version: 1.0.0
  description: Shannon 架构对齐的核心接口与事件流定义
servers:
  - url: /
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/stream/sse:
    get:
      summary: 订阅事件流
      parameters:
        - name: workflow_id
          in: query
          required: true
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: string
        - name: last_event_id
          in: query
          schema:
            type: string
        - name: Last-Event-ID
          in: header
          schema:
            type: string
      responses:
        '200':
          description: text/event-stream
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/StreamEvent'
  /api/v1/events:
    get:
      summary: 查询事件日志
      parameters:
        - name: workflowId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/timeline:
    get:
      summary: 查询时间线
      parameters:
        - name: workflowId
          in: query
          required: true
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [summary, full]
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/v1/timeline/steps:
    get:
      summary: 查询步骤时间线
      parameters:
        - name: workflowId
          in: query
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/StepTimelineResponse'
        '400':
          description: 参数错误，错误码：INVALID_REQUEST、INVALID_CURSOR、TENANT_MISSING
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 访问被拒绝，错误码：POLICY_DENIED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在，错误码：NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 服务不可用，错误码：CIRCUIT_OPEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
        traceId:
          type: string
        requestId:
          type: string
    ErrorResponse:
      description: 错误码命名以 POLICY_DENIED、SANDBOX_DENIED 为准，别名 POLICY_DENY、SANDBOX_DENY 仅用于兼容说明。
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        traceId:
          type: string
        requestId:
          type: string
          description: 请求唯一标识（来自请求上下文/请求头生成），用于端到端排查
...
    StreamEvent:
      type: object
      properties:
        eventId:
          type: string
        schemaVersion:
          type: string
        workflowId:
          type: string
        type:
          type: string
        agentId:
          type: string
        message:
          type: string
        timestamp:
          type: string
        seq:
          type: integer
        streamId:
          type: string
        tenantId:
          type: string
        payload:
          type: object
...
    StepRecord:
      type: object
      properties:
        stepId:
          type: string
        workflowId:
          type: string
        stepSeq:
          type: integer
        type:
          type: string
        status:
          type: string
        attempt:
          type: integer
        input:
          type: object
        output:
          type: object
        errorCode:
          type: string
        tenantId:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
    StepTimelineResponse:
      type: object
      properties:
        workflowId:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/StepRecord'
        nextCursor:
          type: string
        hasMore:
          type: boolean
...